Allow merging into 0-sized piles, even if tags do not match - fixes Inventory give() with nonstackable items
The special case of @count = 0 is used in the Inventory module,
give() method, to properly merge possibly-oversized stacks into
the appropriate stack size of the target. Items with tags cannot
stack (for example, pickaxes), but the tags may not match the
placeholder pile, so just always allow merging into 0-size of same item type.