Fix splitPile(-n)/fraction on Infinity updating pile with count = NaN
Splitting a pile with splitPile(n) where n>0 worked as expected,
but the special forms where n is negative (meaning, get a split
pile with _all but_ -n items), or fractional (for returning the
pile divided into a fractional part, 0.5 = half, etc), would
set the original pile's this.count to NaN because of:

    Infinity - Infinity

resulting in (but not equal to) NaN. Expected result in this case
is Infinity since infinite piles are intended to act as unlimited
sources, even for more infinite piles. To fix this, special-case
n === Infinity and do not perform the subtraction. Also added a
test case.